Mark endless.img padding as allocated via SetFileValidData
SetEndOfFile() allocates extra space as needed, but it does not initialize
it. NTFS guarantees that the extra space is zeroed, but this is actually
implemented in the driver. Files have separate notions of "allocated"
space versus "initialized" space. The driver returns zeroes for reads
beyond the initialized length; and writes zeroes as necessary before
random writes into the allocated-but-uninitialized space.

This leads to the weird situation where the tail of endless.img appears to
be zeroed when read on Windows or with ntfs-3g, even when Endless OS has
written the backup GPT to the last sector returned by ntfsextents. We
suspect that this causes corruption when the disk is defragmented under
Windows: the OS considers the tail of the file to still be uninitialized,
so if its on-disk allocation changes the data which is actually written
there is not preserved.

https://phabricator.endlessm.com/T13309