Install progress: prevent image moving when status is empty
The status is set to the empty string when we move between stages of the
writing process. Because there is now no other significant content in this
container, it collapses and the MarketingContainer below shifts up by one
line-height.

This is kind of an ugly fix but it does work :-)