add /forcembr flag to skip checking for Windows MBR
We found some Samsung recovery partition on a Windows 7 Home laptop which
prevents the installer from proceeding. I don't imagine Samsung shipped a lot
of Win 7 laptops in LatAm, so now that we know it's not a standard Windows MBR,
I don't think it's worth us trying to chase after these Samsung MBRs with a
whitelist. Here's an option which lets the install proceed, as a customer
support option.