show normal first screen when called as endless-uninstaller.exe
We switch the Install button for Uninstall either when invoked as
endless-uninstaller.exe, or if C:\endless\endless.img exists. If the user
cancels the prompt we return to the main screen. This changes the analytics we
send a little, but the key events are unchanged, with the benefit of only
logging screen:UninstallPage and event:Uninstall,Selected if the user says yes
to the prompt. The UninstallFinished pseudo-screen has been removed, when the
session ends during an uninstall it will be the LastPage pseudo-screen.