SetupDualBoot: unset READONLY when using endless.img
On a live USB, the pristine endless.img file has the READONLY flag set.
CopyFileEx propagates it to the installed c:\endless\endless.img. We
ultimately do want that file to be READONLY but first we need to extend
it, and opening it for writing fails if it has the READONLY flag.

We could call GetFileAttributes, mask out FILE_ATTRIBUTE_READONLY, and
then pass the result back to SetFileAttributes but I just don't think
that's necessary: we will later set exactly r|s|h on this file so there's
no harm in setting two of those at the same time as clearing r.

https://phabricator.endlessm.com/T12858