don't turn "cancelled" into download error and only report it once
The user "Cancel" button stops any ongoing downloads and then sends a
"finished all operations" message to stop the other threads. However, if
there is an ongoing download, the cancellation is reported as a download
error, and the code path for download status changes overwrites the
m_lastErrorCause = Cancelled value, and goes to the error page with
"download failed" instead. Handling the operation finished message
then also goes to the error page and re-sends the download failed etc
analytics a 2nd time. So: in the cancellation case, do nothing when
the downloads stop with an "error" and instead leave the operation
finished case to handle and propogate the cancellation "error" in
one place.

https://phabricator.endlessm.com/T13892