Don't hide \endless on combined USBs
If these files are hidden, it's tough for an end user to delete the
Endless OS files from the USB to use it for other purposes, short of
reformatting it or repartitioning it with some other tool (like upstream
Rufus!).

If they can see (and delete) the Endless OS files, of course this is its
own downside, but it's not a total disaster -- there is no persistent data
in there. And this doesn't reclaim *all* space on the disk -- there's
still the ESP and the BIOS boot partition -- but that's a very small
fraction of the disk.

This is also more consistent with eos-write-live-image, which doesn't set
any flags at all! (The exFAT implementation doesn't support them.)