Fetch manufacturer and model from WMI
This API is gross, but it seems to be the best option, and it's endorsed
by Raymond Chen:
https://blogs.msdn.microsoft.com/oldnewthing/20081218-00/?p=19783

https://msdn.microsoft.com/en-us/library/aa394102(v=vs.85).aspx claims
this is a >= Vista API, but it works fine for me on BIOS-booting 32-bit
Windows XP and on EFI-booting 64-bit Windows 10. On XP,
CoInitializeSecurity fails, but the rest of the query still seems to
work fine. We're calling it with exactly the parameters listed on
https://msdn.microsoft.com/en-us/library/aa390423(v=vs.85).aspx so I
don't know why this would fail.

The other API to access this information is GetSystemFirmwareTable;
https://msdn.microsoft.com/en-us/library/ms724379(VS.85).aspx says
this is available in >= Vista or >= XP x64. The fact they explicitly
call out that only the x64 version of XP is supported leads me to
believe that it really wouldn't work on 32-bit XP. Also, it seems to
just give you back a byte array of SMBIOS data, so I think WMI is
actually better.