Refactor "update IE" dialog
The URL we used before points to a page entitled:

> Internet Explorer 7 for Windows XP SP2, Shop With More Confidence Sweepstakes

With the following copy:

> This download is for Microsoft XP Service Pack 2 in English only. This
> custom IE7 download will be available during November 2007 - January
> 31, 2008. This custom download of IE7 will add a homepage tab to
> http://www.microsoft.com/windows/products/winfamily/ie/confidence
> after install.

When I tested on a freshly-created XP VM, I had trouble installing via
this link, but running Windows Update a few times did the trick, so I
think it's better to point to that. Unfortunately the mechanism to
launch the Windows Update UI varies between Windows versions.

<https://msdn.microsoft.com/en-us/library/ms687393(v=vs.85).aspx> says
of WinExec: "If the function succeeds, the return value is greater than
31." It also claims that it's deprecated in favour of CreateProcess but
the MSDN documentation for named control panel pages (link in comment)
uses WinExec so let's roll with it.

https://phabricator.endlessm.com/T14168