Log a metric with system IE version
Our analytics show that 99% of IE users accessing our website have IE8
or later, but of course there may be a group of users who have IE7
installed but use some other modern browser to browse the web.

This event will be logged in the "None" category, I think that's okay.
We could also log an error in this case but it's not actionable and I'd
rather like to keep error events for things we would ideally fix.