Correctly detect IE >= 10
As in so many other cases, it looks like Microsoft had to freeze the major
version of IE at 9, and add another registry key for the true version!

This is harmless since we behave identically for all versions >= 9 at the
moment, but for gathering analytics about our users it's nice to have the
real version.