Analytics: parse debug response, log on error
It's nice to have a line to put a breakpoint on in this case. Note that
the HTTP status code is always 200 regardless of whether the request was
valid or not.

JsonCpp supports parsing from a std::iostream, but I couldn't figure out a
good way to adapt CHttpFile (or, more to the point, its CFile superclass)
to an iostream, so I went via std::string instead. In practice the
response is normally <1KiB.