Analytics: use system IE's User-Agent
As explained in the commentary, this is actually not quite as simple as
it might seem due to the presence of Compatibility View, and we need a
different mechanism to report the *actual* major version of IE in a clear
way.

I think this is okay -- we are mostly interested in the OS version
here, and are using a real UA string because the Google Analytics
documentation contains dire warnings about making up your own UA. Also,
because our analytics ID is for a mobile app, there doesn't actually seem
to be a "Browser Version" field (though I could be overlooking it).