Add Analytics' HANDLE to teardown set
This should help avoid events being lost when the app closes. Also, while
we're here, moving the Closed event here will mean it is reliably sent on
all exit paths. Previously it was only sent when the app was closed with
the [x] / [Finish] button, not when closed with Alt-F4, nor when the
uninstall or IE version dialogs were dismissed.

You might notice that previously, in the timeout case, we would delete the
CWinThread; and in the non-timeout case, we wouldn't. This is because when
the thread terminates, it deletes itself (m_bAutoDelete defaults to TRUE).
It would be a bit fiddly to arrange to manually delete it in the time-out
path here, and AFAICT could lead to a double-free:

    main thread               analytics thread
    -----------               ----------------
    WaitForMultipleObjects
    ... times out
    [pre-empted]              [scheduled]
                              threadSendRequest returns
			      CWinThread deletes itself
    [scheduled]
    delete m_workerThread

(Before I realised m_bAutoDelete == TRUE, I put the 'delete' into
Analytics::~Analytics, which of course crashed on exit!)

https://phabricator.endlessm.com/T13565