SetupDualBoot: create Uninstall reg key sooner
Fully "installing" the uninstaller is now the first action we take. This is
aesthetically pleasing, but more importantly: a 'goto error' *after*
installing GRUB is extremely dangerous. If writing to the registry fails
for any reason, C:\endless will be deleted -- but GRUB will still be
installed, rendering the system unbootable.

The situation where GRUB is installed but C:\endless is missing has been
observed in the wild. This is one potential cause.

https://phabricator.endlessm.com/T14122