WriteMBRAndSBRToWinDrive: make write_bootmark non-fatal
In principle, this should be redundant. We only overwrite the first 440
bytes of the first sector, and Windows has presumably already written the
bootmark at the end of each 512-byte chunk of that first sector, so the
bootmark should already be present.

It is also dangerous to raise a fatal error *after* installing the GRUB
MBR, as this will cause C:\endless to be removed.

https://phabricator.endlessm.com/T14122