SetupDualBoot: check MBR on entry
We found a situation where, due to some state confusion elsewhere in the
application, the user erroneously sees a transient-failure page with a
Resume button, after SetupDualBoot has already run once (installing GRUB).

In this case, we would overwrite the contents of \endless, then attempt to
install GRUB. On BIOS systems, WriteMBRAndSBRToWinDrive would fail (the
"non-Windows MBR" is Endless's GRUB!) and we'd delete \endless, rendering
the system unbootable.

I've taken this opportunity to refactor the up-front checking as well,
since the conditions to be checked are the same.

https://phabricator.endlessm.com/T13599