WM_POWERBROADCAST: fix PBT_APMQUERYSUSPEND check
This is wrong in not one, but two ways.

Firstly, PBT_APMQUERYSUSPEND is a value of wParam, not lParam. (lParam for
PBT_APMQUERYSUSPEND is 0 or 1 depending on whether attempting to deny the
request will actually work.)

Secondly, bitwise AND with zero will always be zero, so this branch would
never be taken:

    #define PBT_APMQUERYSUSPEND             0x0000

In any case, this event is not sent on anything but XP, so ¯\_(ツ)_/¯