Define MSG_045 and MSG_047
These strings are used by Rufus' `GetUSBDevices()` function, which we
use in the USB device selection page.

They are taken verbatim from `res/localization/rufus.loc` so we should
arrange for the existing translations in that file to be imported into
Transifex. Unfortunately it looks like the best way to do that is to do
it by hand.

https://phabricator.endlessm.com/T14144