Make JSON download failures non-fatal
Either we have got past the image selection screen (in which case we don't
care about the JSON), or we have not (in which case that screen has a nice
"internet problems" message we can show, rather than the final, scary
error screen with a misleading message).

Ideally we would retry but I think this is good enough: we allow BITS to
retry for a long time, and we retry whenever the internet connection
flaps.