Simplify DownloadManager::JobError
The block which switches the job to BG_JOB_PRIORITY_FOREGROUND was
redundant:

* Our files are basically always >2GiB
* We always use BG_JOB_PRIORITY_FOREGROUND anyway, for speedier downloads

CComHeapPtr is neater than manually calling CoTaskMemFree.

In practice GetErrorDescription always fails because the message catalogue
in question is not loaded. I do not fully understand how that is meant to
work, but this means that in practice pszErrorDescription is always NULL,
so we need to deal with that.