DownloadManager: report errors from JobError
I've seen cases, apparently random, where JobModification is not called
for BG_JOB_STATE_ERROR. The docs do say that JobError is called from a
higher-priority thread so events may appear out of sequence. I think
what happens is the following race:

    JobModification called
    GetState()
    -> TRANSIENT_ERROR
    [ BITS decides the error is not transient, but fatal: ]
    [ preempted ]             JobError called
                              [ ... does its thing ... ]
                              JobError returns
    [ scheduled ]
    JobModification returns
    [ No state change since the last time JobModification returned, so ]
    [ it is not called again!                                          ]

JobError is guaranteed to be called on error, so let's report errors
from there (rather than it being a no-op) and ignore them in
JobModification. In the process, I simplified JobModification a little
bit.