Respawn UpdateDownloadProgressThread after an error
This thread is only started if m_downloadUpdateThread ==
INVALID_HANDLE_VALUE, but nothing ever set it back to that value if the
thread terminated.

I'm not sure if this is quite right -- are we leaking the handle to the
old thread? -- but it is strictly better than the current behaviour!