Send a "fake" ErrorCauseJSONDownloadFailed; remove it from enum
As of the rest of the refactoring, this is not really an "error cause" any
more -- ErrorCause_t describes reasons why the user hits the ErrorPage.

But to avoid cluttering the analytics with a different representation of
roughly the same event, let's represent this (transient, non-fatal) state
the same way.