Coding: Use CSS to show/hide elements depending on the mode
This calls the setCodingMode() JavaScript function to allow showing/hiding
elements from the HTML UI depending on the operating mode. The function
adds either the "ModeNormal" or "ModeCoding" classes to the <body> element,
and CSS rules which hide the elements nested inside with a non-matching
class. Elements to be shown in both modes do not need any CSS class.