Introduce a "Retrying" state for JSON download
Per https://github.com/endlessm/rufus/pull/59#pullrequestreview-15188058
we want to show the "something's wrong with the connection" message when
there's a transient error fetching the JSON download, but to keep retrying
essentially forever.

In the process of introducing DownloadStatus_t.transientError, arrange for
UpdateDownloadProgressThread/GetDownloadProgress to always fully
initialize the struct.