*Really* correctly use the latest version
The key detail I missed in my previous attempt to fix this bug is that
'persImages' is only assigned to in the first ("pick a version") loop.
This means that even the old logic, which compared the versions as
strings, would actually just pick whichever set of URLs happened to be
last in the JSON file. JsonCpp happens to preserve ordering within
objects, and we happen to sort the keys when generating the JSON, so in
practice this was equivalent for as long as string comparisons on versions
was enough...

So: split the function in two. This narrows the scopes of the variables
such that it doesn't work at all if we don't correctly initialise
persImages.

https://phabricator.endlessm.com/T1451