Generate Version.h from `git describe`
The trick of depending on .git/logs/HEAD comes from
http://neugierig.org/software/blog/2014/11/binary-revisions.html
and I feel like following LLVM's lead is fine.

The restriction on running PowerShell scripts is annoying but maybe we'll
be okay in practice. I took one look at the batch syntax for splitting on
a delimiter and decided I never want to write another batch file.

https://phabricator.endlessm.com/T14955