AppVeyor: verify that rufus.pot has been updated
There is a job which pushes rufus.pot up to Transifex, but it doesn't
run this script (as far as I know). Therefore there's a risk of adding
new strings to endless.loc but not updating rufus.pot, leaving them
untranslated until someone updates it by hand.

So, make updating rufus.pot a condition of the AppVeyor build passing.
Because loc2pot will be run on Windows in this case, we need to take
care to disable Git's newline mangling, and to write the (hopefully
unmodified) rufus.pot using Unix newlines and in UTF-8 to avoid spurious
errors.