ISO: verify endless.img inside endless.squash
For tedious implementation reasons we sign the .img inside the SquashFS
image, not the SquashFS image itself. So, we need to arrange to feed the
uncompressed image through the verification routine, not the outer
SquashFS image.

In principle, of course, we could arrange for the endless.squash file to
be signed. However, in future we will want to be able to install from a
.iso file, distributed as a single file inside a torrent, so we'll need
similar logic to this in future anyway.