Show 'sea' personality as id, th, vi
If we are running in one of those locales, we show it as that language; if
not, we show it as the three languages, comma-separated.

The other option would be to show three options, one for each language, in
the list. That's a bit more invasive to achieve, so I took this route. We
need logic like this for the local image list, in any case, unless we want
to show three entries in *that* list for one file.

With this change, the personality-based sort order appears even more wrong
even in English:

* Arabic
* Bengali
* English
* Spanish
* Spanish (Guatemala)
* Spanish (Mexico)
* French
* Portuguese (Brazilian)
* Indonesian, Thai, Vietnamese
* Chinese (Simplified)

At least it's stable!

https://phabricator.endlessm.com/T15642