Tidy up verifying more than one file
The previous logic involved swapping around which file was pointed to by
m_localFile: it could be the OS image, the boot.zip bundle, or the
eosinstaller image.

b1f7ff8 introduced a regression where we'd try to verify a
newly-downloaded OS image against the signature for the boot bundle. The
offending code was:

    if (verifiedBootFilesZip) {
        // we checked the boot archive signature, now check the image
        m_localFile = UTF8ToCString(image_path);
        GetSignatureForLocalFile(m_localFile, m_localFileSig);

`GetSignatureForLocalFile` would fail: there is no `struct FileImageEntry'
for the newly-downloaded image, so GetSignatureForLocalFile would fail,
and m_localFileSig would still hold the path to the boot.zip.asc.

Instead, we maintain an explicit queue of files to verify, and run the
verification thread once. This cleans up a lot of long-standing TODOs, as
well as fixing the bug at hand.

https://phabricator.endlessm.com/T15658