7zpp: use 7z's CInFileStream to read files
The stream returned by SHCreateStreamOnFileEx does not behave correctly
when reading past the 2**32 byte boundary. Read() is supposed to advance
the cursor by the number of bytes read, but if you cross 2**32 bytes it
instead advances the cursor by 4096 bytes (despite claiming that the
correct number of bytes were read). This, of course, renders the file
apparently corrupt: 7z depends on being able to Read(1) then Seek(-1) to
peek one byte ahead.

7z, of course, already contains an implementation of its own IInStream
type that operates on local files. It's not exported in the normal DLL,
but that is no barrier here: we statically link the library so the world
is our oyster.

https://stackoverflow.com/questions/42399071/shcreatestreamonfileex-on-files-larger-than-232-bytes
https://phabricator.endlessm.com/T14309