GeneralCode: add HRESULT-checking macro
IFFALSE_PRINTERROR(SUCCEEDED(hr), "blahblah") prints the GetLastError()
message, which is misleading because the real error is in 'hr', not
GetLastError().

Add a new set of macros that check a HRESULT and print it on error. We
take care to only evaluate the 'hr' argument once, to allow these macros
to be called with a function call in that position.