Distinguish CantCheckMBR from NonWindowsMBR
NonWindowsMBR is one of the most common errors reported in our
analytics, but I think many of them may be "can't check MBR", not "I
read the MBR and it's not Windows'".

It's also better to show a generic "something went wrong" message, rather
than "you have another OS installed" which is probably not the case.

https://phabricator.endlessm.com/T14868