Skip SelectFilePage if exactly one image is available
In Coding mode, there is only one image (base). If we have no local images
to display, just skip the image selection page entirely.

If there's 0 or >1 remote image, or at least one local image, show the
page. If the manifest has not yet been fetched, we'll show the selection
page, and fill in the Basic half of the pane once it's fetched. This is
not ideal but will do for now.