SetEndlessRegistryKey: simplify
The only remaining case where createBackup=false is the recursive call
within this function. There are no cases where it is called with a string
argument, and UnsetEndlessRegistryKey doesn't implement that case anyway.

So, we can remove the createBackup parameter and the !createBackup
branches of conditionals; remove the string case; and simplify. I am
leaving the switch around the variant for symmetry with
UnsetEndlessRegistryKey (which has the same degenerate switch) and in case
other types actually are needed in future.