Don't check MBR if eosldr is available
When we're about to start installing the OS, we can check whether eosldr
is available; and only if not do we double-check the MBR is what we
expect.

At the front screen of the app, we do not have the boot.zip yet, so cannot
check. In practice, anyone using >=3.2.y.z of this tool will have a new
enough image, either because it's on the same USB stick, or because
they're about to download the latest release. So, we err on the side of
assuming it will be available, and check again when we're about to
install.