Initialize FormatStatus before GetPhysicalHandle
When this is called, we are either not in a cancellable state (the MBR
check on the first page) or have just checked for cancellation.
GetPhysicalHandle depends on this variable being initialized for its
retry loop to work, but as far as I can tell, nothing necessarily
initializes it.

It is conceivable that this is the cause of
https://phabricator.endlessm.com/T14868 -- the user chooses dual-boot
install on the front page; we try to check the MBR but fail to lock the
drive; so the user cannot install.

(cherry picked from commit a2a62f92df4afeab8507cf15ad0ce3d4ca61c81d)