FormatFirstPartitionOnDrive: fix retry loop
FormatDrive sets FormatStatus to non-zero on error, and FormatExCallback
(used within FormatStatus) silently early-returns if
IsError(FormatStatus). FormatDrive does not initialize FormatStatus to 0
on entry. It follows that calling it in a loop without reinitializing
FormatStatus is equivalent to calling it once, since the second and
subsequent calls will all silently fail immediately.

(Contrary to what this comment claims, Rufus itself does not call
FormatDrive in a loop, but hey ho. Other things do get retried in this
manner; maybe this will help us.)

https://phabricator.endlessm.com/T14852