Refuse to run on Windows XP or Vista
These versions are no longer supported by Microsoft.

On XP, the installer cannot contact the Endless OS download server, since
it requires HTTPS and SNI (which the Windows XP TLS stack does not
support). It also doesn't work from an exFAT-formatted USB disk, because
XP does not support exFAT.

On Vista, everything currently works if you have upgraded to IE8, but a
negligible number of users are on this version, so since Microsoft have
given up on it, so will we.

https://phabricator.endlessm.com/T17064