Don't convert target image size to/from GiB
I think the intention was to display neat values in the dropdown list but
it introduces the potential for confusion. AFAICT it also means the
minimum size was (uncompressed size + 1GiB) rounded down to the nearest
GiB, and ditto the maximum; I think it's clearer to keep these precise.

I also removed the maximum size. If you have 2 TiB of free space, why
should we not offer values larger than 128 GiB but smaller than 1.99 TiB?