Don't include download size in maximum install size
If the user wants to make the installed Endess OS image as large as
possible, we need to subtract the download size from the free disk space.

Previously we correctly showed an error if there wasn't enough space for
the download + minimum unpacked image size + a gig of slack on C:; but the
maximum size we offered was the entire disk free space, minus 1GiB.  If
the user selected this option, the download would succeed, the image would
be unpacked, and then extending the image would fail. Users would actually
then be able to go back and install from the pre-downloaded image -- in
which case the maximum size would be chosen correctly -- but let's avoid
that happening at all.

https://phabricator.endlessm.com/T13351
(cherry picked from commit 5a8afc93eb50897e8ce3a5d5e768db52d15b2e9d)