Increase minimum free space for installation to 4GiB
Upgrading across major versions can require more than 1 GiB (the
previous minimum free space); and in the 'base' image, installing any
GNOME applications requires installing the GNOME runtime which is
another gig or so. So: require more space to ensure users don't get
stuck without the ability to upgrade.

Full images ship with the GNOME runtime, but if you have opted for a
full image, you are less concerned about disk space, so for simplicity
we add the same headroom.

https://phabricator.endlessm.com/T17102