endless.loc: correct Mb, Gb -> MB, GB
Regardless of how one defines a gigabyte, the standard abbreviation is
"GB". "Gb" (or, more commonly, "Gbit") means "gigabit": 1 GB = 8 Gbit.

Note that while the IEC recommends that 1 kilobyte, when defined to mean
1000 bytes, should be spelled "1 kB" (with a lowercase "k"), Windows
uses "1 KB" (with an uppercase "K", meaning 1024 bytes). We follow this
convention here, as does Rufus upstream.

https://phabricator.endlessm.com/T12718