BCD: set boot menu timeout to >= 30 seconds
We found a system "in the wild" with timeout 0, even though it had no
other OSes installed. Okay, it was lazy of me to not check and set it
before. Windows' default is 30 seconds, at least on every
newly-installed system I can find, so use that value.

Notice that while the timeout is documented to be a 64-bit unsigned
integer, if you fetch it you get a string, and when you set it you need
to pass a 32-bit signed integer.

https://phabricator.endlessm.com/T17014