Fix moving pages with normal vs. code titles
To allow the user to drag the dialog around the screen by its
"titlebar", mousedown events on elements with class PageHeader or
PageHeaderTitle are caught and converted into WM_NCLBUTTONDOWN messages.
The mousedown events are bound to a method using the DHTML_EVENT_CLASS
macro.  Apparently this is implemented in terms of matching the entire
class name, not individual elements of the space-separated class list,
so an element with class="PageHeaderTitle ModeNormal" is not matched by
DHTML_EVENT_CLASS(..., L"PageHeaderTitle", ...).

So, instead, wrap the PageHeaderTitle elements in an extra <div> just to
carry the ModeNormal/ModeCoding class. I tried the opposite way round:

    <div class="PageHeaderTitle">
        <div class="ModeNormal" id="DualBootPageTitle">Endless OS</div>
        <div class="ModeCoding" id="DualBootPageTitleCoding">Endless Coding</div>
    </div>

which looks neater, but also doesn't work, presumably because the event
is fired at the inner element rather than the wrapper .PageHeaderTitle
element. In modern browsers you can use style="pointer-events: none;" to
make an element transparent to clicks but this is not supported on most
of the IE versions we support.

https://phabricator.endlessm.com/T14033