Fetch newer boot.zip as needed for eosldr.
To use eosldr, there's actually no hard dependency on an Endless OS 3.2
image: you just need a new enough boot.zip. So we hardcoded the name and
URL of a newer boot.zip. When scanning local files, if an image is too
old we look for this boot.zip rather than the corresponding one; when
reading the JSON manifest, we substitute this boot.zip as needed.

The net result is that all of these cases work:

* No images downloaded yet => download 3.1.8 image + eos3.2 boot.zip
* 3.1.8 image and boot.zip already downloaded => only download eos3.2
  boot.zip
* 3.1.8 image and hardcoded boot.zip already downloaded => use them.

https://phabricator.endlessm.com/T17244