Simplify FileDescription
This is the string that shows up in the UAC dialog, labelled "Program
name", alongside the "Verified publisher" ("Endless Mobile, Inc", from
our signing cert) and the "File origin" (eg "Hard drive on this
computer"). This long name looks bad in the UAC dialog, and "Endless
Installer" is what we show on the taskbar once the app starts, so I
think we should show the same here.