Fully fork FormatDrive()
If you peer through the #ifdef soup, this is all there is to it. We will
make further changes so that it accepts a path rather than formatting
the first basic data partition on the drive, so best make a clean break.

Rather annoyingly, FormatEx accepts non-const strings as parameters.
Fine: we can copy them. (In practice Rufus itself passes literals for these
two parameters at one call site so it would probably be okay to just cast
away the const.)