Address first partition by full device path
Previously we only knew how to format (essentially) the first basic data
partition on a drive. This meant that we created USB sticks with a FAT32
ESP and an exFAT image partition by creating a partition table that puts
the ESP logically first with partition type basic data, formatting and
populating it, then rearranging the partition table to put the image
partition logically first (fixing the ESP's partition type at the same
time), formatting and populating that.

It's possible to call FormatEx() on any device you can get a path to.  One
way to do this, used by Rufus's SetupWinToGo(), is to use AltMountVolume()
to assign it a drive letter (which looks up a path for it by finding a
path for an existing mounted volume, mapping that to a form where the disk
and partition number are separate, then changing the partition number and
mounting the result).

This seems like a roundabout approach: you can quite easily construct a
path given the drive and partition index. It seems a bit too good to be
true, so we do a fair bit of consistency checking to make sure we are
addressing the correct device.

This will allow us to avoid the partition table switcheroo.