Avoid USB partition layout switcheroo
There is no need to perform this juggling act: if you can address a
drive, you can format it. We know how to address a non-basic-data
partition on a disk, even though it doesn't get assigned a
Volume{GUID}-style path (on Windows versions prior to 1703 "Creators
Update", that is...).

Previously, on error we would unmount whichever partition was mounted at
the time; on success, eoslive would remain mounted. Instead, we now
unconditionally unmount the ESP at the end of the process, and leave the
image partition mounted (if indeed it is).