Write BIOS boot partition before partitioning drive
There's some rationale in the comment. Some further rationale is that
Rufus actually does the same thing when writing its NTFS-UEFI partition (a
VFAT partition containing a UEFI bootloader that loads an NTFS driver,
then chainloads the real bootloader from an NTFS partition).

I believe that writing to the MBR causes Windows to rescan the drive, so
we continue to do that just before we are about to sleep and wait for
the logical volumes to re-appear.