UEFIBootSetup: upper-case hex digits in BootXXXX
From the UEFI Specification, Version 2.7 Errata A, § 3.1.1 Boot Manager
Programming:

> Each load option entry resides in a Boot####, Driver####,
> SysPrep####, OsRecovery#### or PlatformRecovery#### variable where
> #### is replaced by a unique option number in printable hexadecimal
> representation using the digits 0–9, and the upper case versions of
> the characters A–F (0000–FFFF).

Previously, we used lower-case versions of the characters a-f, and would
conclude that Boot000a (for example) is unused, write our boot entry
there, and move the integer 0x000a to the front of BootOrder. This leads
to the existing Boot000A being tried, not our Boot000a.

https://phabricator.endlessm.com/T20920