Only report MaximumInstallSizeGB if it's positive
I noticed that we have a fair few MaximumInstallSizeGB with values like
17,179,869,178 GB (~ 16 EB). This is due to unsigned integer underflow
when downloadSize > (freeBytesAvailable - bytesInGig).

To resolve this, let's only send MaximumInstallSizeGB when there is enough
space to proceed with the download and install. In the case where there is
not enough free space, we now report the free space to the nearest GB
together with that event.

https://phabricator.endlessm.com/T19201