Add GetSelectedRemoteImage() helper
As well as reducing duplicated code, in StartInstallationProcess() we
did not check the result of FindIndex() before passing it to GetAt(). If
the former returns NULL, the latter crashes. We saw this assertion
failing in a debug build; we don't have a backtrace but it seems likely
this is the cause. (It's a should-not-happen code path but there we go.)

The new helper always checks the returned POSITION, and we now handle
this should-not-happen case in StartInstallationProcess() in a similar
way to the local case above it.

https://phabricator.endlessm.com/T20744