WMI: check ConversionStatus as well as ProtectionStatus
Previously, we only checked ProtectionStatus, but as explained in the
code it's possible for the disk to be encrypted even if ProtectionStatus
== 0.

Rather than including the new condition in the query, let's always fetch
the record for the drive letter in question, and explicitly print the
properties we're checking. It would be nice to print all its properties
but today's not the day I write a generic function to print a VARIANT's
type and value. (Although perhaps _variant_t's implicit conversion to
string would do it?)

https://phabricator.endlessm.com/T21417